requirements.txt
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.3.1
djangorestframework==3.16.1
gunicorn==23.0.0
joblib==1.3.2
kombu==5.5.4
numpy==1.24.3
packaging==25.0
pandas==2.1.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.0.0
scikit-learn==1.3.2
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
